feat : service unit test 추가 및 레이어 리팩토링

+++ src/controllers/users.controller.js
@@ -39,16 +39,22 @@ export class UsersController {
         password
       );
 
-      // 쿠키에 accessToken 설정
-      res.cookie("accessToken", `Bearer ${accessToken}`);
-      console.log(accessToken);
+      //accessToken과 refreshToken을 쿠키에 설정
+      res.cookie("accessToken", accessToken, {
+        httpOnly: true,
+        secure: true,
+      });
+      res.cookie("refreshToken", refreshToken, {
+        httpOnly: true,
+        secure: true,
+      });
 
       return res.status(200).send({
         message: "로그인 되었습니다",
-        // data: {
-        //   accessToken,
-        //   refreshToken,
-        // },
+        data: {
+          accessToken,
+          refreshToken,
+        },
       });
     } catch (err) {
       console.error(err);
@@ -56,19 +62,31 @@ export class UsersController {
     }
   };
 
+  // 로그아웃
   signout = async (req, res, next) => {
     try {
-      const accessToken = req.cookies.accessToken;
+      // 쿠키에서 accessToken과 refreshToken 제거
+      const clearAccessToken = await this.usersService.clearAccessTokenCookie(
+        res
+      );
+      const clearRefreshToken = await this.usersService.clearRefreshTokenCookie(
+        res
+      );
 
-      if (!accessToken) {
-        // 클라이언트에게 로그인 상태가 아니라는 메시지 전달
+      // 쿠키 제거 중 오류가 발생하면 처리
+      if (!clearAccessToken || !clearRefreshToken) {
+        return res
+          .status(500)
+          .json({ message: "쿠키 제거 중 오류가 발생했습니다." });
+      }
+
+      // 클라이언트에게 로그인 상태가 아니라는 메시지 전달
+      if (!req.cookies.accessToken || !req.cookies.refreshToken) {
         return res.status(401).json({ message: "로그인되어 있지 않습니다." });
       }
 
+      // 로그아웃 메서드 호출
       const result = await this.usersService.signout();
-
-      // 쿠키에서 accessToken 제거
-      res.clearCookie("accessToken", { path: "/", secure: true });
 
       return res.status(200).json(result);
     } catch (err) {
@@ -76,11 +94,22 @@ export class UsersController {
     }
   };
 
-  // 회원 탈퇴
+  // 계정삭제
   deleteUser = async (req, res, next) => {
     try {
-      const { userId } = req.params;
+      const userId = req.params.id;
+      const { password } = req.body;
 
+      // 사용자 비밀번호 검증
+      const isPasswordValid = await this.usersService.verifyUserPassword(
+        userId,
+        password
+      );
+      if (!isPasswordValid) {
+        return res
+          .status(401)
+          .json({ message: "비밀번호가 올바르지 않습니다." });
+      }
       const deleteUser = await this.usersService.deleteUser(userId);
 
       return res.status(200).json({ data: deleteUser });
